pyproject.toml
[tool.poetry]
name = "switchboard"
authors = ["Atilio Barreda/D4BL <atilio@miski.studio>"]
readme = "README.md"
package-mode = false
packages = [{include = "pre_processing", from = "."}]

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
fpdf = "^1.7.2"
ocrmypdf = "^16.5.0"
pymupdf = "^1.24.10"
spacy = "^3.7.6"
boto3 = "^1.35.39"
requests-aws4auth = "^1.3.1"
elasticsearch = "^8.15.1"
uvicorn = "^0.31.1"
pandas = "^2.2.3"
pdfminer-six = "^20240706"
jupyterlab = "^4.2.5"
bertopic = "^0.15.0"
sentence-transformers = "^2.2.2"
umap-learn = "^0.5.3"
hdbscan = "^0.8.29"
plotly = "^5.14.1"
networkx = "^3.4.2"
pyvis = "^0.3.2"
pypdf2 = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

